Fix eslint linting errors and separate from to its own component
