+++ src/pages/Home.js
@@ -1,5 +1,12 @@
+import { Fragment } from 'react';
 import AllCities from '../containers/AllCities';
+import NavBar from '../containers/NavBar';
 
-const Home = () => <AllCities />;
+const Home = () => (
+  <>
+    <NavBar renderedComponent="all-cities" />
+    <AllCities />
+  </>
+);
 
 export default Home;
